Improve Output Writing Logic
1. Remove unnecessary setup_time_steps() function

This function basically overrode the SimulationVariables setup.
It did not change the time stamp and the current time step size, so this part was irrelevant.
Additionally, it set max_simulation_time to the current time step size.
This is counterintuitive and did not seem to be helpful in any way.

2. Add correct input variable to run_timestep() calls

In ceval.cpp, when calling the timesteppers, a time stamp should be parsed
(apparently this is necessary for dealing with tidal waves).
In the previous implementation, max_simulation_time was parsed here.
I changed it to the current_simulation_time, as it should be
(at least when considering the function description).

3. Only output the last SDC iteration by default

WIP: This works but should be a command line argument in the future.
I did this to avoid getting eight output files for every time step - we are usually only interested in the final iteration anyways.

4. Switch to a Different Outputting Logic

Introduced a new function that checks whether output is necessary or not: timestep_check_output().
This function is inspired by the function with the same name in swe_sphere.cpp.
However, I use a bit simpler logic - check out the comments for details.

TODO here: how should we treat the corner case where we are in the last time step in the simulation?
Should both cfinal() and cecho_output_solution() write output?
Or only one of them?
Do we always want output written from cfinal()?